# [nomerge] faster ++ and union for HashSet