… src/library/scala/collection/immutable/HashSet.scala