for 2.12 only - this will not merge directly into 2.13. A separate PR will be needed if any of these changes need to be ported to 2.13

faster operations (mostly union and ++) for HashSet
EmptySet and SetBuilder adjusted to take advantage of ++ optimisation for simple and common cases

eliminate unneeded allocations for HashSets where the result is already built for -
   subSet ++ superSet
   subSet union superSet
   superSet union subSet
   superSet ++ subSet

make a fast path when there is structural sharing in the HashSet
for union, guarantee internal operations will only return a new HashSet if one of the existing HashSet parameters or internal values cant be used

use System.arraycopy rather than Array.copy as it avoid JVM nulling the array

add missing `eq` fast path to intersect0 and diff0

minor improvements to + to reduce allocations

reduce calls to size which is expensive, especially when we know the size beforehand