@mkeskells 2019-12-12
performance result in https://gist.github.com/mkeskells/a3ec74d3dd79c6564ac957c8387c8fda

based on these benchmarks this PR is generally faster, for this ++ 
for `opContainedWithLarge` (sharing makes no difference for this test) **roughly a 10x speedup**
this is roughly `subset ++ superset`
before 
```
[info] HashSetPlusPlusBenchmark.opContainedWithLarge          0      10  avgt   20       9744.980 ▒      142.660  ns/op
[info] HashSetPlusPlusBenchmark.opContainedWithLarge          0     100  avgt   20     151431.353 ▒     9977.708  ns/op
[info] HashSetPlusPlusBenchmark.opContainedWithLarge          0    1000  avgt   20    1809340.500 ▒    20265.960  ns/op
[info] HashSetPlusPlusBenchmark.opContainedWithLarge          0   10000  avgt   20   24871168.850 ▒  1172582.403  ns/op
[info] HashSetPlusPlusBenchmark.opContainedWithLarge          0  100000  avgt   20  307419271.995 ▒ 11660918.310  ns/op```
after
```
[info] HashSetPlusPlusBenchmark.opContainedWithLarge          0      10  avgt   20        422.985 ▒       3.649  ns/op
[info] HashSetPlusPlusBenchmark.opContainedWithLarge          0     100  avgt   20      11030.800 ▒     305.684  ns/op
[info] HashSetPlusPlusBenchmark.opContainedWithLarge          0    1000  avgt   20      97682.239 ▒    9048.045  ns/op
[info] HashSetPlusPlusBenchmark.opContainedWithLarge          0   10000  avgt   20    1659434.234 ▒   27740.896  ns/op
[info] HashSetPlusPlusBenchmark.opContainedWithLarge          0  100000  avgt   20   27916736.500 ▒ 2031136.722  ns/op
```
for `opEmptyWithContained` ( roughly `HashSet.empty ++ set`) the result is constant time
again sharing doesnt mean anything
before
```
[info] HashSetPlusPlusBenchmark.opEmptyWithContained          0      10  avgt   20       7974.439 ▒      736.207  ns/op
[info] HashSetPlusPlusBenchmark.opEmptyWithContained          0     100  avgt   20     159306.679 ▒     9042.004  ns/op
[info] HashSetPlusPlusBenchmark.opEmptyWithContained          0    1000  avgt   20    1786056.237 ▒    10988.317  ns/op
[info] HashSetPlusPlusBenchmark.opEmptyWithContained          0   10000  avgt   20   24018732.300 ▒   559415.648  ns/op
[info] HashSetPlusPlusBenchmark.opEmptyWithContained          0  100000  avgt   20  304566590.500 ▒  6853616.865  ns/op
```
after
```
[info] HashSetPlusPlusBenchmark.opEmptyWithContained          0      10  avgt   20          6.147 ▒       0.325  ns/op
[info] HashSetPlusPlusBenchmark.opEmptyWithContained          0     100  avgt   20          5.943 ▒       0.089  ns/op
[info] HashSetPlusPlusBenchmark.opEmptyWithContained          0    1000  avgt   20          6.082 ▒       0.208  ns/op
[info] HashSetPlusPlusBenchmark.opEmptyWithContained          0   10000  avgt   20          5.993 ▒       0.106  ns/op
[info] HashSetPlusPlusBenchmark.opEmptyWithContained          0  100000  avgt   20          5.925 ▒       0.052  ns/op
```
for `opLargeWithContained` ( roughly `superset ++ subset)`, where the subset has the same values, but no structural sharing, is slower, **more than 50% slower**
before 
```
[info] HashSetPlusPlusBenchmark.opLargeWithContained          0      10  avgt   20        359.570 ▒        2.689  ns/op
[info] HashSetPlusPlusBenchmark.opLargeWithContained          0     100  avgt   20       5847.542 ▒      339.311  ns/op
[info] HashSetPlusPlusBenchmark.opLargeWithContained          0    1000  avgt   20      84791.967 ▒     6336.078  ns/op
[info] HashSetPlusPlusBenchmark.opLargeWithContained          0   10000  avgt   20    1463687.727 ▒    70435.294  ns/op
[info] HashSetPlusPlusBenchmark.opLargeWithContained          0  100000  avgt   20   17498114.320 ▒  2309792.760  ns/op
```
after
```
[info] HashSetPlusPlusBenchmark.opLargeWithContained          0      10  avgt   20        417.390 ▒       4.413  ns/op
[info] HashSetPlusPlusBenchmark.opLargeWithContained          0     100  avgt   20       9560.103 ▒      71.380  ns/op
[info] HashSetPlusPlusBenchmark.opLargeWithContained          0    1000  avgt   20      90247.421 ▒    3834.973  ns/op
[info] HashSetPlusPlusBenchmark.opLargeWithContained          0   10000  avgt   20    1617386.403 ▒   27404.463  ns/op
[info] HashSetPlusPlusBenchmark.opLargeWithContained          0  100000  avgt   20   28008580.875 ▒ 1885487.745  ns/op
```
`opLargeWithEmpty` is trivial in both cases
`opWithDistinct` - adding non overlapping sets is about **3x faster**
before 
```
[info] HashSetPlusPlusBenchmark.opWithDistinct                0      10  avgt   20        673.877 ▒        5.234  ns/op
[info] HashSetPlusPlusBenchmark.opWithDistinct                0     100  avgt   20      13343.080 ▒      428.135  ns/op
[info] HashSetPlusPlusBenchmark.opWithDistinct                0    1000  avgt   20     175657.800 ▒     1678.009  ns/op
[info] HashSetPlusPlusBenchmark.opWithDistinct                0   10000  avgt   20    2175801.267 ▒     7979.342  ns/op
[info] HashSetPlusPlusBenchmark.opWithDistinct                0  100000  avgt   20   27818889.500 ▒   270538.247  ns/op
```
after
```
[info] HashSetPlusPlusBenchmark.opWithDistinct                0      10  avgt   20        211.629 ▒       1.629  ns/op
[info] HashSetPlusPlusBenchmark.opWithDistinct                0     100  avgt   20       2963.875 ▒       9.130  ns/op
[info] HashSetPlusPlusBenchmark.opWithDistinct                0    1000  avgt   20      53410.397 ▒    2963.410  ns/op
[info] HashSetPlusPlusBenchmark.opWithDistinct                0   10000  avgt   20     668198.521 ▒   16270.055  ns/op
[info] HashSetPlusPlusBenchmark.opWithDistinct                0  100000  avgt   20    9144742.953 ▒  850943.160  ns/op
```
for `opWithOverlap` adding 2 set that partially overlap, but don't share structure **2-3 x faster**
before
```
[info] HashSetPlusPlusBenchmark.opWithOverlap                 0      10  avgt   20       2767.593 ▒      160.711  ns/op
[info] HashSetPlusPlusBenchmark.opWithOverlap                 0     100  avgt   20      70057.189 ▒     2797.278  ns/op
[info] HashSetPlusPlusBenchmark.opWithOverlap                 0    1000  avgt   20    1008365.163 ▒     3398.963  ns/op
[info] HashSetPlusPlusBenchmark.opWithOverlap                 0   10000  avgt   20   10091560.364 ▒    73586.293  ns/op
[info] HashSetPlusPlusBenchmark.opWithOverlap                 0  100000  avgt   20  109812473.000 ▒  7878452.591  ns/op
```
after
```
[info] HashSetPlusPlusBenchmark.opWithOverlap                 0      10  avgt   20       1629.035 ▒       8.210  ns/op
[info] HashSetPlusPlusBenchmark.opWithOverlap                 0     100  avgt   20      21293.519 ▒     533.269  ns/op
[info] HashSetPlusPlusBenchmark.opWithOverlap                 0    1000  avgt   20     298748.476 ▒   10874.936  ns/op
[info] HashSetPlusPlusBenchmark.opWithOverlap                 0   10000  avgt   20    3936182.795 ▒   22880.808  ns/op
[info] HashSetPlusPlusBenchmark.opWithOverlap                 0  100000  avgt   20   38748065.167 ▒  882330.327  ns/op
```
`opWithShared` is similar to `opWithOverlap`, but with shared structure. Here the degree of structural sharing makes a difference. in general is seems to be **3-4 x faster**
```
[info] Benchmark                                      (sharing)  (size)  Mode  Cnt          Score          Error  Units
[info] HashSetPlusPlusBenchmark.opWithShared                  0      10  avgt   20      11797.167 ▒       86.355  ns/op
[info] HashSetPlusPlusBenchmark.opWithShared                  0     100  avgt   20     200615.259 ▒     4293.505  ns/op
[info] HashSetPlusPlusBenchmark.opWithShared                  0    1000  avgt   20    2311063.558 ▒   220145.374  ns/op
[info] HashSetPlusPlusBenchmark.opWithShared                  0   10000  avgt   20   29573128.750 ▒   122618.359  ns/op
[info] HashSetPlusPlusBenchmark.opWithShared                  0  100000  avgt   20  401452987.000 ▒ 14649242.141  ns/op
[info] HashSetPlusPlusBenchmark.opWithShared                 20  100000  avgt   20  333808538.495 ▒  4499989.419  ns/op
[info] HashSetPlusPlusBenchmark.opWithShared                 40  100000  avgt   20  286483932.505 ▒  6256244.707  ns/op
[info] HashSetPlusPlusBenchmark.opWithShared                 60  100000  avgt   20  230013784.000 ▒  3560713.801  ns/op
[info] HashSetPlusPlusBenchmark.opWithShared                 80  100000  avgt   20  165039797.000 ▒  3532941.437  ns/op
[info] HashSetPlusPlusBenchmark.opWithShared                 90  100000  avgt   20  150678145.000 ▒  5145690.234  ns/op
[info] HashSetPlusPlusBenchmark.opWithShared                100  100000  avgt   20  110482294.250 ▒  7684039.614  ns/op
```
after
```
[info] Benchmark                                      (sharing)  (size)  Mode  Cnt          Score          Error  Units
[info] HashSetPlusPlusBenchmark.opWithShared                  0      10  avgt   20       2883.413 ▒      36.261  ns/op
[info] HashSetPlusPlusBenchmark.opWithShared                  0     100  avgt   20      52635.662 ▒     117.919  ns/op
[info] HashSetPlusPlusBenchmark.opWithShared                  0    1000  avgt   20     541960.551 ▒   25599.479  ns/op
[info] HashSetPlusPlusBenchmark.opWithShared                  0   10000  avgt   20    9532744.977 ▒  415512.332  ns/op
[info] HashSetPlusPlusBenchmark.opWithShared                  0  100000  avgt   20  116030047.000 ▒ 9916748.064  ns/op
[info] HashSetPlusPlusBenchmark.opWithShared                 20  100000  avgt   20  107303203.750 ▒ 8584760.299  ns/op
[info] HashSetPlusPlusBenchmark.opWithShared                 40  100000  avgt   20   95388560.750 ▒ 6409158.068  ns/op
[info] HashSetPlusPlusBenchmark.opWithShared                 60  100000  avgt   20   76985525.500 ▒ 3989717.795  ns/op
[info] HashSetPlusPlusBenchmark.opWithShared                 80  100000  avgt   20   60160765.500 ▒ 3734495.243  ns/op
[info] HashSetPlusPlusBenchmark.opWithShared                 90  100000  avgt   20   51376904.500 ▒ 2859621.033  ns/op
[info] HashSetPlusPlusBenchmark.opWithShared                100  100000  avgt   20          6.031 ▒       0.172  ns/op
```