@mkeskells 2020-01-23
@szeiger I have some, but from running on a laptop and some are less stable then I hoped
@rorygraves is getting a better set ATM. To run the full dataset takes 36 hours for the baseline, and another 36 for the post PR

the results and a comparison are 
https://1drv.ms/x/s!Am3TAKFvNVy_m1xzYxrldZKiY_Jx?e=zIidmA

On sheet 3 - change the dataset and the operation to see the effect
not all combinations are valid - the unshared benchmark only has one of the data points, because they dont have the degree of how much is shared

I am less sure on how to visualise the data, what we have is a range of different sizes and datasets, and operations, and CPU/allocation to compare

the operations tested are 

'+'
'--'
'-'
'++'
'--'
'union'
'diff'
'intersect'
'subsetOf'
'sameElements'
 
on datasets of 
Shared benchmarks - 
**opWithOverlapShared** - operations on data with structural sharing
**opWithOverlapUnshared** - operation on data with shared values, but not shared HashSet structure

both of the above operate for different amounts of overlap

**opContainedSharedWithData** - subset with a superset (e.g. smallSet ++ superset), with structural sharing
**opContainedUnsharedWithData** - subset with a superset (e.g. smallSet ++ superset), without structural sharing, just shared values
**opDataWithContainedShared** - superset with subset (e.g. superset ++ smallSet ), with structural sharing 
**opDataWithContainedUnshared** - superset with subset (e.g. superset ++ smallSet ), without structural sharing 
**opDataWithEmpty** - e.g. set ++ HashSet.empty
**opDataWithSetEmpty** -  e.g. set ++ Set.empty
**opEmptyWithData** - e.g. HashSet.empty ++ set
**opSetEmptyWithData** - Set.empty ++ set
**opWithDistinct** - set ++ otherSet

The tests allow to data that heavily collides on hashcode , or doesn't, but I don't have the data for colliding data yet

so for ++, opWithOverlapShared data  the results for different degrees of sharing are  

-- | -- | -- | -- | -- | -- | --
  |   | size | 10 | 100 | 1000 | 10000
  |   | old NS | 13,370 | 188,566 | 2,176,863 | 31,044,478
  |   | new NS | 3,028 | 49,987 | 487,436 | 9,043,898
  |   | old all | 26,374 | 377,310 | 4,729,831 | 59,706,222
  |   | new all | 2,653 | 27,777 | 228,820 | 2,590,108
  |   |   |   |   |   |  
  |   |   |   |   |   |  
  | duration | improvement | 77.35% | 73.49% | 77.61% | 70.87%
  | allocation | improvement | 89.94% | 92.64% | 95.16% | 95.66%
  |   |   |   |   |   |  
  |   |   |   |   |   |  
  |   |   |   |   |   |  
  |   |   |   |   |   |  
  |   | size | 10 | 100 | 1000 | 10000
  |   | old NS | 12,758 | 164,534 | 1,873,404 | 28,038,742
  |   | new NS | 2,799 | 46,087 | 462,166 | 8,294,675
  |   | old all | 20,862 | 299,673 | 3,760,636 | 47,639,448
  |   | new all | 2,346 | 21,873 | 186,546 | 2,235,285
  |   |   |   |   |   |  
  |   |   |   |   |   |  
  | duration | improvement | 78.06% | 71.99% | 75.33% | 70.42%
  | allocation | improvement | 88.75% | 92.70% | 95.04% | 95.31%
  |   |   |   |   |   |  
  |   |   |   |   |   |  
  |   |   |   |   |   |  
  |   |   |   |   |   |  
  |   | size | 10 | 100 | 1000 | 10000
  |   | old NS | 9,699 | 143,708 | 1,564,424 | 23,910,666
  |   | new NS | 2,402 | 41,425 | 419,419 | 7,498,449
  |   | old all | 15,478 | 222,370 | 2,801,492 | 35,628,647
  |   | new all | 1,975 | 16,034 | 149,434 | 1,810,038
  |   |   |   |   |   |  
  |   |   |   |   |   |  
  | duration | improvement | 75.23% | 71.17% | 73.19% | 68.64%
  | allocation | improvement | 87.24% | 92.79% | 94.67% | 94.92%
  |   |   |   |   |   |  
  |   |   |   |   |   |  
  |   |   |   |   |   |  
  |   |   |   |   |   |  
  |   | size | 10 | 100 | 1000 | 10000
  |   | old NS | 8,141 | 111,515 | 1,213,250 | 19,219,547
  |   | new NS | 2,126 | 34,545 | 350,257 | 6,563,930
  |   | old all | 10,204 | 146,391 | 1,854,356 | 23,679,103
  |   | new all | 1,391 | 10,474 | 116,687 | 1,270,059
  |   |   |   |   |   |  
  |   |   |   |   |   |  
  | duration | improvement | 73.88% | 69.02% | 71.13% | 65.85%
  | allocation | improvement | 86.37% | 92.85% | 93.71% | 94.64%
  |   |   |   |   |   |  
  |   |   |   |   |   |  
  |   |   |   |   |   |  
  |   |   |   |   |   |  
  |   | size | 10 | 100 | 1000 | 10000
  |   | old NS | 6,022 | 82,897 | 874,119 | 13,959,189
  |   | new NS | 1,537 | 25,691 | 273,990 | 4,693,990
  |   | old all | 5,042 | 72,263 | 919,776 | 11,797,190
  |   | new all | 665 | 6,094 | 76,314 | 619,438
  |   |   |   |   |   |  
  |   |   |   |   |   |  
  | duration | improvement | 74.48% | 69.01% | 68.66% | 66.37%
  | allocation | improvement | 86.82% | 91.57% | 91.70% | 94.75%
  |   |   |   |   |   |  
  |   |   |   |   |   |  
  |   |   |   |   |   |  
  |   |   |   |   |   |  
  |   | size | 10 | 100 | 1000 | 10000
  |   | old NS | 3,903 | 74,473 | 926,667 | 18,337,603
  |   | new NS | 9 | 9 | 9 | 9
  |   | old all | 40 | 40 | 42 | 68
  |   | new all | 0 | 0 | 0 | 0
  |   |   |   |   |   |  
  |   |   |   |   |   |  
  | duration | improvement | 99.78% | 99.99% | 100.00% | 100.00%
  | allocation | improvement | 100.00% | 100.00% | 100.00% | 100.00%

all of the ++ operation look good to me, with one exception dataWithContainedShared/unshared
both show a blip on the wall time for size 100 case that I have not had time to look at

  |   | size | 10 | 100 | 1000 | 10000
-- | -- | -- | -- | -- | -- | --
  |   | old NS | 333 | 5,218 | 122,479 | 1,481,221
  |   | new NS | 217 | 6,672 | 65,459 | 762,688
  |   | old all | 40 | 40 | 40 | 42
  |   | new all | 0 | 0 | 0 | 1
  |   |   |   |   |   |  
  |   |   |   |   |   |  
  | duration | improvement | 34.88% | -27.87% | 46.55% | 48.51%
  | allocation | improvement | 100.00% | 99.97% | 99.73% | 97.02%


generally there is a marked improvement for '++' sameElements, 
measureable improvement for 'union', 
not much difference for '-', '--', intersect

but there are some anomalies e.g. for diff on large sets

e.g.

dataset | scala.collection.immutable.HashSetBulkSharedBenchmark.opWithOverlapShared
-- | --
operation | diff
  |   |   |   |   |   |   |   |  
  |   |   |   | size | 10 | 100 | 1000 | 10000
dataset | scala.collection.immutable.HashSetBulkSharedBenchmark.opWithOverlapShared | old NS | 2,275 | 37,495 | 431,304 | 5,654,334
  |   |   |   | new NS | 2,359 | 38,364 | 451,010 | 7,090,971
colliding | FALSE |   |   | old all | 1,132 | 6,408 | 62,070 | 585,699
operation | diff |   |   | new all | 1,091 | 6,026 | 57,986 | 565,573
sharing | 0 |   |   |   |   |   |   |  
  |   |   |   |   |   |   |   |  
  |   | duration | improvement | -3.71% | -2.32% | -4.57% | -25.41%
  |   | allocation | improvement | 3.60% | 5.95% | 6.58% | 3.44%